Expression textbox now shows up on filter icon click
