+++ src/js/components/pipelines/DataTable.jsx
@@ -13,9 +13,11 @@ var d3 = require('d3'),
     HoverValue = require('./HoverValue');
 
 function mapStateToProps(state, ownProps) {
+  var id = ownProps.id;
   return {
-    dataset: getInVis(state, 'datasets.' + ownProps.id),
-    vega: state.get('vega')
+    dataset: getInVis(state, 'datasets.' + id),
+    vega: state.get('vega'),
+    expressionTextbox: getInVis(state, 'datasets.' + id + '._expressionTextbox')
   };
 }
 
@@ -89,7 +91,15 @@ var DataTable = React.createClass({
         keys = dl.keys(schema),
         max = output.length,
         fmt = dl.format.auto.number(),
-        scrollLeft = this.$table && this.$table.node().scrollLeft;
+        scrollLeft = this.$table && this.$table.node().scrollLeft,
+        expTextbox = props.expressionTextbox,
+        expTextbox = expTextbox ? expTextbox.toJS() : expTextbox,
+        showTextbox = expTextbox && expTextbox.show;
+
+    var textbox = showTextbox ? (
+      <input type="text" name="expression"
+        placeholder="Enter Expression" />
+    ): null;
 
     var prev = page > 0 ? (
       <Icon glyph={assets.prev} width="10" height="10" onClick={this.prevPage} />
@@ -101,6 +111,9 @@ var DataTable = React.createClass({
 
     return (
       <div>
+
+        {textbox}
+
         <div className="datatable"
           onMouseLeave={this.hideHover} onScroll={this.hideHover}>
